Изменил абстрактный класс и наследование
